refactor: move pennylane dep fxns back to workflow
